+++ common/src/main/java/net/smackplays/smacksutil/items/LightWand.java
@@ -25,7 +25,7 @@ public class LightWand extends Item {
         BlockPos pos = context.getClickedPos();
         Level world = context.getLevel();
         Player player = context.getPlayer();
-        if (world.isClientSide) return super.useOn(context);
+        if (world.isClientSide) return InteractionResult.SUCCESS;
         BlockPos toPlace = pos.relative(face, 1);
 
         if (world.getBlockState(toPlace).is(Blocks.AIR) && player != null && !player.isCrouching()) {
@@ -33,6 +33,7 @@ public class LightWand extends Item {
             if (!player.isCreative()) {
                 context.getItemInHand().hurt(1, player.getRandom(), (ServerPlayer) player);
             }
+            return InteractionResult.SUCCESS;
         }
         return super.useOn(context);
     }