+++ neoforge/src/main/java/net/smackplays/smacksutil/networking/ServerEnchantPayloadHandler.java
@@ -18,11 +18,11 @@ public class ServerEnchantPayloadHandler {
         // Do something with the data, on the main thread
         context.workHandler().submitAsync(() -> {
                     ItemStack stack = data.item();
-                    Player player = context.player().get();
-                    if (player != null){
+                    if (context.player().isPresent()){
+                        Player player = context.player().get();
                         AbstractContainerMenu containerMenu = player.containerMenu;
                         containerMenu.slots.get(0).set(stack);
                     }
-                });
+        });
     }
 }